services/lyrics-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
numpy==1.26.4
openai-whisper==20231117
torch==2.1.2
soundfile==0.12.1